Fix Firefox display issue
When the stereo component runs, Firefox looks for the cubemap mesh but
cannot find it causing the variable (object3D) to be undefined. Chrome,
on the other hand, does find it so no error occurs.

Temporary fix is to run a short script at the end of the page that adds
the stereo component after the scene has loaded (ensuring each entity
has a mesh).